+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to
 
 ### Added
 
+- Implement edx video browser events pydantic models
 - Create a `post` endpoint for statements implementing the LRS spec
 - Implement support for the MongoDB database backend
 - Implement support for custom queries when using database backends `get`